+++ src/netbox_contract/tables.py
@@ -100,6 +100,9 @@ class ContractListTable(NetBoxTable):
     external_partie_object = tables.Column(verbose_name='External partie', linkify=True)
     parent = tables.Column(linkify=True)
     yrc = tables.Column(verbose_name='Yerly recuring costs')
+    status = columns.ChoiceFieldColumn(
+        verbose_name=('Status'),
+    )
 
     class Meta(NetBoxTable.Meta):
         model = Contract
@@ -118,7 +121,6 @@ class ContractListTable(NetBoxTable):
             'initial_term',
             'renewal_term',
             'currency',
-            'accounting_dimensions',
             'mrc',
             'yrc',
             'nrc',
@@ -173,7 +175,6 @@ class InvoiceListTable(NetBoxTable):
             'period_start',
             'period_end',
             'currency',
-            'accounting_dimensions',
             'amount',
             'documents',
             'comments',
@@ -225,6 +226,9 @@ class InvoiceLineListTable(NetBoxTable):
 
 
 class AccountingDimensionListTable(NetBoxTable):
+    status = columns.ChoiceFieldColumn(
+        verbose_name=('Status'),
+    )
     class Meta(NetBoxTable.Meta):
         model = AccountingDimension
         fields = (
@@ -232,9 +236,11 @@ class AccountingDimensionListTable(NetBoxTable):
             'name',
             'value',
             'comments',
+            'status',
         )
         default_columns = (
             'name',
             'value',
             'comments',
+            'status',
         )